feat: Add support for max downloads, exclude files, and duplicate handling
The download function in google-drive-utils has been updated to support the following features:
- Download a maximum number of files
- Exclude specific files from the download
- Handle duplicate filenames by skipping, overwriting, or renaming the file

In addition, the directory structure has been reorganized to put the CLI functionality in a corresponding submodule, and the entry point for the CLI scripts has been changed.

BREAKING CHANGES:
- The signature of the download function has changed to add new parameters for max downloads, exclude files, and duplicate handling. If you are using the download function in an existing codebase, you may need to update your code to use the new parameters.
- If you were using the CLI scripts in the previous version of google-drive-utils, you will need to update your command to use the new entry point.

This update improves the flexibility and usability of google-drive-utils, making it easier to download files from Google Drive with more control over the download process.